credit_risk_fyp/requirements.txt
# Core ML Libraries
numpy==1.23.5
pandas==2.0.3
scikit-learn==1.3.0

# Deep Learning (GPU-enabled)
tensorflow-gpu==2.13.0
tensorflow-addons==0.21.0
keras==2.13.0

# Gradient Boosting Libraries (GPU support)
xgboost==2.0.0
lightgbm==4.0.0
catboost==1.2

# Data Processing
dask[complete]==2023.5.0
joblib==1.3.0
pyarrow==12.0.0

# Imbalanced Learning
imbalanced-learn==0.11.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Model Interpretability
shap==0.42.0

# Progress Bars
tqdm==4.65.0

# Configuration
pyyaml==6.0
python-dotenv==1.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Utilities
colorama==0.4.6
